requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
coverage==7.2.2
Django==4.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2023.2
sqlparse==0.4.3
tomli==2.0.1
